Add hipSYCL support for sycl-acc and sycl-usm w/ device-aware MPI Split update_tile_halo_kernel/update_halo into multiple TUs for faster compile time

+++ test.sh
@@ -7,13 +7,31 @@ export CUDA_DIR="$NVHPC_DIR/cuda/"
 export PATH=$NVHPC_DIR/compilers/bin/:${PATH:-}
 export KOKKOS_DIR="/home/tom/Downloads/kokkos-4.0.01/"
 
+export HIPSYCL_DEBUG_LEVEL=1
+
 export CPU_RANKS=16
 export GPU_RANKS=3
 
 VERBOSE="ON"
 PROBLEM="InputDecks/clover_bm_short.in"
 
+declare -A enabled_models=(
+  ["serial"]=true
+  ["omp"]=true
+  ["omp-target"]=true
+  ["kokkos"]=true
+  ["cuda"]=true
+  ["hip"]=true
+  ["std-indices"]=true
+  ["sycl-acc"]=true
+  ["sycl-usm"]=true
+)
+
 function test() {
+  if [[ "${enabled_models[$3]}" != true ]]; then
+    echo "# Skipping model $3"
+    return
+  fi
   rm -rf "build_$3" # /CMakeCache.txt
   echo "$2" "${@:4}"
   cmake -DCMAKE_BUILD_TYPE=Release -S. -B "build_$3" -GNinja -DCMAKE_VERBOSE_MAKEFILE=$VERBOSE -DENABLE_MPI=ON -DMODEL="$3" "${@:4}" # &>/dev/null
@@ -34,11 +52,14 @@ function test() {
 }
 
 function test_nompi() {
+  if [[ "${enabled_models[$2]}" != true ]]; then
+    echo "# Skipping model $2"
+    return
+  fi
   rm -rf "build_$2" # /CMakeCache.txt
   echo "$2" "${@:3}"
   cmake -DCMAKE_BUILD_TYPE=Release -S. -B "build_$2" -GNinja -DCMAKE_VERBOSE_MAKEFILE=$VERBOSE -DENABLE_MPI=OFF -DMODEL="$2" "${@:3}" # &>/dev/null
   cmake --build "build_$2"                                                                                                            # &>/dev/null
-
   export OMP_TARGET_OFFLOAD=MANDATORY
   export OMP_PLACES=cores
   export OMP_PROC_BIND=true
@@ -49,11 +70,12 @@ function test_nompi() {
   build_$2/*-cloverleaf --device "$1" --file "$PROBLEM" --profile --staging-buffer auto
 }
 
-(
-
-  test_nompi "0" serial -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=g++
-  test_nompi "0" serial -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=clang++
-
+( 
+  (
+    :
+    test_nompi "0" serial -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=g++
+    test_nompi "0" serial -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=clang++
+  )
   (
     :
     test_nompi "0" std-indices -DCMAKE_CXX_COMPILER=g++ -DCXX_EXTRA_FLAGS="-Ofast" -DUSE_TBB=ON
@@ -92,9 +114,60 @@ function test_nompi() {
     test_nompi "0" hip -DMANAGED_ALLOC=ON -DCXX_EXTRA_FLAGS="-O1" -DCMAKE_CXX_COMPILER=hipcc # fails validation at > O1 lol
     test_nompi "0" hip -DSYNC_ALL_KERNELS=ON -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=hipcc
   )
+  (
+    :
+    export HIPSYCL_TARGETS="omp.library-only"
+    test_nompi "OpenMP" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test_nompi "OpenMP" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HIPSYCL_TARGETS="cuda:sm_60"
+    test_nompi "NVIDIA" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test_nompi "NVIDIA" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HSA_OVERRIDE_GFX_VERSION=10.1.2 #  https://github.com/RadeonOpenCompute/ROCm/issues/1756#issuecomment-1160386571
+    export HIPSYCL_TARGETS="hip:gfx1012"
+    test_nompi "Radeon" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test_nompi "Radeon" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+  )
+  (
+    :
+    set +eu
+    module load mpi
+    source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
+    source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh
+    set -eu
+    export DPCPP_CPU_NUM_CUS=1
+    export DPCPP_CPU_SCHEDULE=static
+    test_nompi "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=ON
+    test_nompi "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
+    test_nompi "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=ON
+    test_nompi "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
+  )
+  (
+    :
+    set +eu
+    module load mpi
+    source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
+    source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh --include-intel-llvm
+    set -eu
+    cuda_sycl_flags="-fsycl-targets=nvptx64-nvidia-cuda;-Xsycl-target-backend;--cuda-gpu-arch=sm_60;--cuda-path=$CUDA_DIR"
+    test_nompi "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$cuda_sycl_flags"
+    test_nompi "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test_nompi "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test_nompi "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test_nompi "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+
+    export LD_LIBRARY_PATH=/opt/rocm-5.4.3/lib:${LD_LIBRARY_PATH:-}
+    hip_sycl_flags="-fsycl;-fsycl-targets=amdgcn-amd-amdhsa;-Xsycl-target-backend;--offload-arch=gfx1012"
+    test_nompi "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$hip_sycl_flags"
+    test_nompi "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test_nompi "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test_nompi "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test_nompi "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+  )
   wait
 )
+
 #exit 0
+
 (
   module load mpi
   test $CPU_RANKS "0" serial -DCMAKE_CXX_COMPILER=g++
@@ -135,44 +208,59 @@ function test_nompi() {
     test $GPU_RANKS "0" hip -DMANAGED_ALLOC=ON -DCXX_EXTRA_FLAGS="-O1" -DCMAKE_CXX_COMPILER=hipcc # fails validation at > O1 lol
     test $GPU_RANKS "0" hip -DMANAGED_ALLOC=ON -DSYNC_ALL_KERNELS=ON -DCXX_EXTRA_FLAGS="-O1" -DCMAKE_CXX_COMPILER=hipcc
   )
+  (
+    :
+    set +eu
+    module load mpi
+    source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
+    source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh
+    set -eu
+    export DPCPP_CPU_NUM_CUS=1
+    export DPCPP_CPU_SCHEDULE=static
+    test $CPU_RANKS "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=ON
+    test $CPU_RANKS "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
+    test $CPU_RANKS "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=ON
+    test $CPU_RANKS "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
+  )
+  (
+    :
+    export HIPSYCL_TARGETS="omp.library-only"
+    test $GPU_RANKS "OpenMP" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "OpenMP" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HIPSYCL_TARGETS="cuda:sm_60"
+    test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HSA_OVERRIDE_GFX_VERSION=10.1.2 #  https://github.com/RadeonOpenCompute/ROCm/issues/1756#issuecomment-1160386571
+    export HIPSYCL_TARGETS="hip:gfx1012"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+  )
+  (
+    :
+    set +eu
+    module load mpi
+    source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
+    source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh --include-intel-llvm
+    set -eu
+    cuda_sycl_flags="-fsycl-targets=nvptx64-nvidia-cuda;-Xsycl-target-backend;--cuda-gpu-arch=sm_60;--cuda-path=$CUDA_DIR"
+    test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$cuda_sycl_flags"
+    test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+
+    export LD_LIBRARY_PATH=/opt/rocm-5.4.3/lib:${LD_LIBRARY_PATH:-}
+    hip_sycl_flags="-fsycl;-fsycl-targets=amdgcn-amd-amdhsa;-Xsycl-target-backend;--offload-arch=gfx1012"
+    test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+  )
   wait
 )
-#exit 0
-#
-(
-  set +eu
-  module load mpi
-  source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
-  source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh
-  set -eu
-  export DPCPP_CPU_NUM_CUS=1
-  export DPCPP_CPU_SCHEDULE=static
-  test $CPU_RANKS "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
-  test $CPU_RANKS "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
-)
 
-(
-  set +eu
-  module load mpi
-  source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
-  source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh --include-intel-llvm
-  set -eu
-
-  cuda_sycl_flags="-fsycl-targets=nvptx64-nvidia-cuda;-Xsycl-target-backend;--cuda-gpu-arch=sm_60;--cuda-path=$CUDA_DIR"
-  test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$cuda_sycl_flags"
-  test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
-  test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
-  test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
-  test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
-
-  export LD_LIBRARY_PATH=/opt/rocm-5.4.3/lib:${LD_LIBRARY_PATH:-}
-  hip_sycl_flags="-fsycl;-fsycl-targets=amdgcn-amd-amdhsa;-Xsycl-target-backend;--offload-arch=gfx1012"
-  test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$hip_sycl_flags"
-  test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-  test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-  test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-  test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-)
+# exit 0
 
 ### CUDA-aware MPI ###
 (
@@ -190,7 +278,16 @@ function test_nompi() {
     #    test $GPU_RANKS "0" hip  -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=/usr/lib/aomp_17.0-1/bin/hipcc # doesn't work with NVHPC'S OpenMPI, segfaults at runtime
 
   )
-
+  (
+    :
+    export HIPSYCL_TARGETS="cuda:sm_60"
+    test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HSA_OVERRIDE_GFX_VERSION=10.1.2 #  https://github.com/RadeonOpenCompute/ROCm/issues/1756#issuecomment-1160386571
+    export HIPSYCL_TARGETS="hip:gfx1012"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+  )
   (
     :
     set +eu
